requirements.txt
tensorflow==2.13.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.13.0
matplotlib
numpy